rename some code
