--- backend/src/romanizer.py
@@ -1,32 +1,0 @@
-import pykakasi
-
-def jap2romaji(original:str) -> str:
-    kakasi = pykakasi.kakasi()
-    romajiMorphemeList = kakasi.convert(original.replace(" ",""))
-
-    romajiSentence = ""
-    for romajiMorpheme in romajiMorphemeList:
-        romajiSentence += romajiMorpheme["hepburn"] + " "
-    return romajiSentence
-
-if __name__=="__main__":
-    sampleText = """
-    私の妻は霊感があるようで、付き合い始めてからなぜか心霊体験することが多くなり、これはそのうちの一つです。 
-    私の妻は釣りが趣味で、私はあまり得意ではありませんが付き合いでよく夜釣りに行っていました。
-    青森県の太平洋側に住んでおり、近くに釣りのポイントもたくさんあるのですが、たまには違うところで釣りをしたい、という妻の意見から日本海側まで車を走らせることにしました。 
-    釣り雑誌を見ながら釣りたい魚を決め、その魚がよく釣れる、という漁港までひたすら走っていましたが、近づくにつれて雨風は強くなり、目的の漁港に着いたときには強風と土砂降りで、立っているだけで体力が持っていかれるような状況でした。 
-    他にも釣り目当てで来ているような車は2台ほど停まっていましたが、どちらも車内で天候が回復するのを待っているようです。 
-    私たち夫婦もしばらく待ってみましたが、天気は良くなるどころか、さらに強さを増す一方で、帰ることも考えましたが、片道5時間かけて来たこともあり、このまま帰るのはもったいない、と少しだけ釣ることにしました。 
-    妻は慣れた手つきでライフジャケットを着けると、釣り道具を持ち、遠くまで歩いて行ってしまいました。 あまり慣れていない私は、車のすぐそばで準備をして釣り糸を垂らしたのですが、風が強すぎて当たりが全くわからず、ただ餌を取られるだけでした。 
-    それから天気の悪い中釣りを続けていると、何を言っているかは分かりませんが、女性の声が聞こえ、そちらを向くと妻が歩いて来ていました。 やっぱりこの天気だから諦めようと言っているんだな、と解釈し、釣竿を片付けて車に戻りました。 
-    二人で車に乗り込み、その旅行を後にした後の夫婦の会話が以下の通りです。 やっぱり釣れなかったね。こんな天気だとは思わなかった。天気予報と全然違うし。仕方ないよ、初めて来る場所だし。そっちが声かけてこなかったらもうちょいやってたかも。 
-    え?声なんてかけてないけど。 は?よく聞こえなかったけど、なんか言ってたじゃん。 
-    妻の話では、防波堤の先で釣りをしていたが全く当たりがなく、ポイントを変えようか迷っていた時に、男性の声が聞こえ、近くに誰もいなかったことから、私が妻に向かって帰ろうということを叫んでいるのだ、と思ったそうです。 
-    確かに私が釣りをしていたところと妻が釣りをしていたところは、漁港の入口付近と端でだいぶ距離が離れていました。 冷静に考えれば、防風雨の中でよく聞き取れないとはいえ、声が聞こえるのもおかしい話です。 
-    その事実に気づいた後、家に着くまで会話らしい会話もなく、ただひたすら車を走らせるだけでした。 家に着いた時には日が昇り朝になっていましたが、眠気はなくとりあえずリビングのテレビをつけました。 
-    地元の情報番組が流れる中で、二人とも釘付けになったニュースがありました。 私たちが釣りに行った漁港の沖で漁を行っていた漁船が転覆し、70代の夫婦が亡くなった、というニュースです。 
-    私たちが聞いた声は、魂だけになった夫婦だったのでしょうか。 未だに答えは出ませんが、その日以降、夜釣りに行くことはなくなりました。
-    """.replace("\n","")
-
-    romaji = jap2romaji(sampleText)
-    print(romaji)
