+++ backend/utils/listener.py
@@ -1,7 +1,7 @@
+import os
+
 import openai
 from dotenv import load_dotenv
-
-import os
 
 load_dotenv(verbose=True)
 openai.api_key = os.getenv("OPENAI_API_KEY")
@@ -11,9 +11,8 @@ audiopath = os.path.join(basedir, "../audio/")
 
 def callWhisper(fname:str) -> str:
     fpath = os.path.join(audiopath, fname)
-
-    audio_file= open(fpath, "rb")
-    transcript = openai.Audio.transcribe("whisper-1", audio_file)
+    with open(fpath, "rb") as audio_file:
+        transcript = openai.Audio.transcribe("whisper-1", audio_file)
     return transcript.text
 
 if __name__=="__main__":
